dte_stand/dte_stand/algorithm/mate/requirements.txt
absl-py==0.9.0
appnope==0.1.0
astor==0.8.1
astroid==2.5.7
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cached-property==1.5.2
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
clang==5.0
cloudpickle==1.6.0
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
dm-tree==0.1.5
entrypoints==0.3
flatbuffers==1.12
gast==0.3.3
gin-config==0.3.0
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
ignnition==1.0.2
importlib==1.0.4
importlib-metadata==1.6.0
importlib-resources==5.1.4
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
notebook==6.0.3
numpy==1.18.3
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.4
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
pylint==2.8.2
pyparsing==2.4.6
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.4.1
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
tensorboard==2.3.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorboardX==1.9
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-probability==0.11.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.9
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
